Update proptype for textStyle
Styles can be an array of style objects or object. Array gives an effect similar to css.

Previously, this value was set to Text.propTypes.style which would allow arrays to be passed without warning in <1.4.0. After upgrading package, I now this see this warning as I am passing [defaultStyle, themeStyle] as a prop to this component.

This will remove this warning.